Import the new kernel contributed by ARM using the new dot-product instructions (UDOT) available on newer CPUs such as Cortex-A76.
This particular kernel is not tuned for one particular CPU; it does
well on Cortex-A76 in particular.  ARM had contributed another kernel
more specifically optimized for Cortex-A55r1; it would have to be
imported separately.

Context: for ARM's contributions, see
https://github.com/google/gemmlowp/pull/116

Notice that the new kernel is not automatically enabled when the
instructions are present: the user is required to also define
a preprocessor token, GEMMLOWP_DOTPROD_KERNEL, to opt in to using
that kernel. Rationale: avoid worsening the ODR-violations situations
with more inline symbols having different definitions based on
preprocessor tokens. Conveniently, this will also allow in the future
to have multiple such kernels coexist, behind separate opt-in tokens.